[Refactor] Take time-of-flight instead of ticks
Attempting to perform timing calculations in McStas (C) using seconds
and ticks was proving troublesom. Instead move those calculations into
the C++ code with a new class to handle the details.

The calling component should pass the time-of-flight which the Readout
would normally truncate to time since the last pulse; but since the
current simulations lack any primary spectrometer the 'full' time of
flight is passed along instead for now.